fix(pulumi-aws): extract APW into apps/api while in development
