+++ packages/pulumi-aws/src/apps/index.ts
@@ -1,4 +1,5 @@
 export * from "./core";
+export * from "./common";
 export * from "./api";
 export * from "./admin";
 export * from "./website";
